requirements.txt
certifi==2021.10.8
cffi==1.15.1
configparser==4.0.2
contextlib2==0.6.0.post1
coverage==5.5
cryptography==3.3.2
enum34==1.1.10
idna==2.10
importlib-metadata==2.1.3
ipaddress==1.0.23
numpy==1.16.6
pandas==0.24.2
pathlib2==2.3.7.post1
pkg-resources==0.0.0
psycopg2-binary==2.8.6
pycparser==2.21
pyOpenSSL==21.0.0
python-dateutil==2.8.2
pytz==2022.2.1
scandir==1.10.0
selenium==3.141.0
six==1.16.0
SQLAlchemy==1.4.40
sqlparse==0.3.1
typing==3.10.0.0
urllib3==1.26.12
urllib3-secure-extra==0.1.0
zipp==1.2.0
